第5刷での修正

+++ 第7章/リスト7-12.cs
@@ -2,7 +2,7 @@
 class BmiCalculator
 {
     // 身長はcm単位で、体重はkg単位で渡してもらう
-    public double GetBmi(double height, double weight)
+    public double GetBmi(int height, int weight)
     {
         var metersTall = height / 100.0;
         var bmi = weight / (metersTall * metersTall);
